feat(osmrBackend): entities, services
Integrate nearest point functionality and make basic osmrbackend service abstract to recycle functionality
